+++ packages/ragbits-core/CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
+
+- Feat: added support for IVFFlat indexing and Halfvec datatype
+
 - Added Lazy loading of dependencies in local.py and during importing of LiteLLM
 - Add tool_choice parameter to LLM interface (#738)
 - Fix Prompt consumes same iterator twice leading to no data added to chat (#768)
+
 
 ## 1.2.2 (2025-08-08)